+++ examples/show_in_mujoco.py
@@ -43,7 +43,7 @@ if __name__ == "__main__":
     try:
         model = load_robot_description(args.name)
     except ModuleNotFoundError:
-        model = load_robot_description(f"{args.name}_description")
+        model = load_robot_description(f"{args.name}_mj_description")
 
     data = mujoco.MjData(model)
     viewer = mujoco_viewer.MujocoViewer(model, data)